powerpc/fsl: p2020si-post: Reroute PCIe@9000 INTB to MPIC 0
Workaround for MVME2500: the board's VME bridge wants to present its
VIRQ on MPIC input 0, but the p2020si-post PCIe@9000 interrupt-map
claims MPIC 9 for INTB. Remap INTB to MPIC 0 so the VME bridge can
own the line.

This is a hack: the change also affects other p2020-based boards
sharing p2020si-post.dtsi, but none of them are built in this tree
in practice.

Signed-off-by: Alessio Igor Bogani <alessio.bogani@elettra.eu>